Remove testify dependency
Remove testify dependency to reduce dependency vulnerability risk.
Because this project is not expected to be actively developed on
it would be best to remove as many dependencies as possible to avoid
extra work that may come up to keep dependencies up to date for security
reasons.

By removing testify we also remove several transitive dependencies.